Табличный ввод услуг и приложения

+++ views/client/addPrice.php
@@ -17,11 +17,11 @@ use yii\helpers\Html;use yii\widgets\ActiveForm; ?>
     <?= $form->field($model, 'PRICE')->textInput(['autocomplete'=>'off'])->label('Цена')?>
 
     <div class="row">
-        <div class="col-md-2"><?= Html::submitButton('Отправить',['class'=>'btn btn-success'])?></div>
+        <div class="col-md-2"><?= Html::submitButton('Сохранить',['class'=>'btn btn-success'])?></div>
 
        <?php if ($_GET['ID_PR']!=NULL):?>
         <div class="col-md-10" style="text-align: right">
-            <a href="index.php?r=client/pricedelete&ID_PR=<?=$model->ID_PR?>" class="btn btn-danger" >Удалить</a>
+            <a href="index.php?r=client/pricedelete&ID_PR=<?=$model->ID_PR?>" class="btn btn-danger" onclick='return confirm("Вы уверены?")' >Удалить</a>
         </div>
         <?php endif;?>
     </div>
